techpack: display: Avoid shutting off HBM when display is off
Every time the display is turned off, mdss-dsi-hbm-off-command is called
to shut off HBM. However, this is sometimes called after
mdss-dsi-post-off-command is called, resulting in a deadlock that keeps
the display black.

To prevent this, check if HBM is enabled first. If it is not, skip
sending the command entirely.

Fixes https://gitlab.com/LineageOS/issues/android/-/issues/4790

Change-Id: Ic6ffc052a594304bff05f64c6fadad147da24c61